backend/requirements.txt
Flask
Flask-SocketIO
Flask-JWT-Extended
Flask-Login
flask-cors
python-dotenv
